Olympics_Project/requirements.txt
flask
pandas
psycopg2-binary
sqlalchemy